requirements.txt
django==3.2.14
djangorestframework==3.13.1
psycopg2-binary==2.9.3
celery[redis]
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-environ==0.9.0
beautifulsoup4==4.11.1
requests
gunicorn
gevent